Mov IK methods into own class
