+++ Visuo-haptic Toolkit/Assets/Visuo-Haptic Toolkit/Scripts/Utils/Gradient.cs
@@ -6,7 +6,7 @@ using UnityEngine.U2D.IK;
 using UnityEngine.UIElements;
 
 namespace VHToolkit {
-    static class MathTools {
+    public static class MathTools {
         /// <summary>
         /// Compute the gradient of its input <c>function</c> at position <c>x</c> by the central differences method.
         /// </summary>
@@ -58,34 +58,6 @@ namespace VHToolkit {
         static public Vector2 ProjectToHorizontalPlane(this Vector3 v) => new(v.x, v.z);
         static public Vector3 LiftFromHorizontalPlane(this Vector2 v) => new(v.x, 0, v.y);
 
-        // Compute the forward kinematics chain from link-to-link length and angle information.
-        static public List<PositionAndRotation2D> ForwardKinematics(PositionAndRotation2D origin, List<float> lengths, List<float> angles) {
-            Debug.Assert(lengths.Count == angles.Count);
-            List<PositionAndRotation2D> result = new() { origin };
-            var position = origin.position;
-            var direction = origin.forward.LiftFromHorizontalPlane();
-            foreach (var (length, angle) in lengths.Zip(angles)) {
-                Debug.Assert(length > 0);
-                Debug.Assert(Mathf.Abs(angle) <= 180f);
-                position += length * result.Last().forward;
-                direction = Quaternion.AngleAxis(angle, Vector3.up) * direction;
-                result.Add(new(position, direction.ProjectToHorizontalPlane()));
-            }
-            return result;
-        }
 
-        static public List<Vector2> InverseKinematics(Vector2 origin, Vector2 end, List<float> lengths) {
-            // TODO figure out sensible solver limit and tolerance
-            bool wasSolved = false;
-            int solverLimit = 1;
-            float tolerance = Vector2.kEpsilon;
-            Vector2[] positions = new Vector2[lengths.Count - 1];
-            while (!wasSolved) {
-                wasSolved = FABRIK2D.Solve(end, solverLimit, tolerance, lengths.ToArray(), ref positions);
-                solverLimit *= 2;
-                tolerance *= 2;
-            }
-            return positions.Select(v => origin + v).ToList();
-        }
     }
 }
