+++ Visuo-haptic Toolkit/Assets/Visuo-Haptic Toolkit/Scripts/Utils/InverseKinematics.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 497bf484155385d42a5d09aeea5e5b8d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
